exposing API + version bump
