+++ core/src/main/scala/gwi/saturator/DagMock.scala
@@ -1,44 +1,8 @@
 package gwi.saturator
 
-import scala.math.Ordering
-
-trait DagPartition {
-  def pid: String
-}
-
-trait DagVertex {
-  def vid: String
-}
-
-object DagVertex {
-  protected[saturator] object State {
-    val Pending = "Pending"
-    val Complete = "Complete"
-    val InProgress = "InProgress"
-    val Failed = "Failed"
-  }
-}
-case class Dependency(p: DagPartition, sourceVertices: Set[DagVertex], targetVertex: DagVertex)
-
-object Dependency {
-  implicit def dependencyOrdering(implicit po: Ordering[DagPartition], vo: Ordering[DagVertex]) = new Ordering[Dependency] {
-    override def compare(x: Dependency, y: Dependency): Int = {
-      val partitionOrd = po.compare(x.p,y.p)
-      lazy val sourceVertexOrd = vo.compare(x.sourceVertices.head, y.sourceVertices.head)
-      lazy val targetVertexOrd = vo.compare(x.targetVertex, y.targetVertex)
-      if (partitionOrd != 0)
-        partitionOrd
-      else if (sourceVertexOrd != 0)
-        sourceVertexOrd
-      else
-        targetVertexOrd
-    }
-  }
-
-}
-
 /**
-  * !!! Test environment !!! To build Dag using just Ints instead of Vertices and Longs instead of Partitions
+  * !!! Test environment !!! it is needed in core tests and example projects to, "test->compile" dependency is awful
+  * To build Dag using just Ints instead of Vertices and Longs instead of Partitions
   */
 protected[saturator] object DagMock {
   import scala.language.implicitConversions
